+++ crates/bin/src/lib.rs
@@ -1,11 +1,13 @@
 #![cfg_attr(docsrs, feature(doc_auto_cfg))]
 
-pub mod args;
-pub mod bin_util;
-pub mod entry;
+mod args;
+mod bin_util;
+mod entry;
 mod git_credentials;
 mod install_path;
-pub mod logging;
-mod manifests;
+mod logging;
+mod main_impl;
 mod signal;
 mod ui;
+
+pub use main_impl::do_main;
